pyproject.toml
[tool.poetry]
name = "trading-platform"
version = "0.1.0"
description = ""
authors = ["Kate Zinchenko <id.a1430@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.2.5"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.8"
pytz = "^2023.3.post1"
sqlparse = "^0.4.4"
PyYAML = "^6.0.1"
typing-extensions = "^4.8.0"
tzdata = "^2023.3"
wrapt = "^1.15.0"
asgiref = "^3.7.2"
astroid = "^2.15.8"
cfgv = "^3.4.0"
colorama = "^0.4.6"
dill = "^0.3.7"
distlib = "^0.3.7"
filelock = "^3.12.4"
identify = "^2.5.30"
isort = "^5.12.0"
lazy-object-proxy = "^1.9.0"
mccabe = "^0.7.0"
nodeenv = "^1.8.0"
platformdirs = "^3.10.0"
pre-commit = "^3.4.0"
python-dotenv = "^1.0.0"
tomli = "^2.0.1"
tomlkit = "^0.12.1"
virtualenv = "^20.24.5"
psycopg = "^3.1.12"
pyjwt = "^2.8.0"
celery = "^5.3.4"
redis = "^5.0.1"
mock = "^5.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.isort]
line_length = 100

[tool.mypy]
